fix(common,trade): use BigDecimal for portfolio weights to prevent precision loss
Replace f64 with BigDecimal for weight fields in TradingAction variants
(Rebalance, AddPosition, ReducePosition) and PortfolioWeights.weights.

Internal optimization calculations (ndarray) remain Vec<f64>; conversion
to BigDecimal happens at the boundary via weight_from_f64() with 10-digit
rounding. This prevents precision loss when weights flow into NearValue
multiplication for trade amount calculations.

+++ crates/common/src/algorithm/types.rs
@@ -153,12 +153,18 @@ pub enum TradingAction {
     },
     /// ポートフォリオリバランス
     Rebalance {
-        target_weights: BTreeMap<TokenOutAccount, f64>,
+        target_weights: BTreeMap<TokenOutAccount, BigDecimal>,
     },
     /// ポジション追加
-    AddPosition { token: TokenOutAccount, weight: f64 },
+    AddPosition {
+        token: TokenOutAccount,
+        weight: BigDecimal,
+    },
     /// ポジション削減
-    ReducePosition { token: TokenOutAccount, weight: f64 },
+    ReducePosition {
+        token: TokenOutAccount,
+        weight: BigDecimal,
+    },
 }
 
 // ==================== アルゴリズム実行結果 ====================
@@ -232,7 +238,7 @@ impl ExecutionReport {
 /// ポートフォリオの重み
 #[derive(Debug, Clone, Serialize, Deserialize)]
 pub struct PortfolioWeights {
-    pub weights: BTreeMap<TokenOutAccount, f64>,
+    pub weights: BTreeMap<TokenOutAccount, BigDecimal>,
     pub timestamp: DateTime<Utc>,
     pub expected_return: f64,
     pub expected_volatility: f64,
